evaluation/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ora-evaluation"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client==2.151.0",
    "google-auth==2.30.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "gspread==6.1.2",
    "python-dotenv==1.0.1",
    "requests==2.32.4",
    "streamlit==1.37.0",
    "arize-phoenix-evals>=2.10.0",
    "langchain-google-vertexai==2.0.15",
    "asyncio==3.4.3",
    "huggingface-hub==0.26.2",
    "instructor[vertexai]==1.5.2",
    "openai==1.58.1",
    "protobuf>=5.29.6",
    "pydantic==2.10.4",
    "tqdm==4.67.1",
    "plotly==5.24.1",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

[project.optional-dependencies]
test = [
    "mypy==1.10.1",
    "ruff==0.5.1",
    "types-requests==2.32.0.20250602",
    "google-api-python-client-stubs==1.28.0",
    "types-tqdm==4.67.0.20241221",
]

[tool.hatch.build.targets.wheel]
packages = ["auto_evaluation", "human_evaluation", "script_based_evaluation"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
disable_error_code = ["call-arg"]
explicit_package_bases = true
exclude = ["src/post_install.py", "auto_evaluation/deprecated"]

[[tool.mypy.overrides]]
module = ["huggingface_hub.*", "vertexai.*", "pandas.*", "plotly.*", "phoenix.*", "google.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["human_evaluation.*"]
ignore_errors = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9","E301","E304","E305","E401","E223","E224","E242", "E", "F" ,"N", "W", "C90"]
extend-select = ["D204"]
ignore = ["E501", "C901", "D203"]
preview = true

# Allow fix for all enabled rules (when `--fix` is provided).
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
